Add a utility to conditionally submit parallel tasks to ForkJoinPool
Added `ParallelExecutor` utility to submit parallel stream tasks
to the `PhysicalCoreExecutor` only if the caller thread is not a `ForkJoinPool` worker thread.

If the caller is already inside a `ForkJoinPool`, then task will be run on the caller thread directly.
That way, parallelization of the streams (index building or compressed vector computation) can be
defined dynamically by the caller `ForkJoinPool`.