Parallelize GraphIndexBuilder.removeDeletedNodes()
`GraphIndexBuilder.addNNDescentConnections()` was being called serially for each
effected node during cleanup after some nodes are marked as deleted.

When there are too many (like hundreds of thousands), `GraphIndexBuilder.cleanup()`
was taking too much time, and most of it was in `removeDeletedNodes()` while calling
`addNNDescentConnections()` in loop.

With this change `addNNDescentConnections()` is called in parallel using SIMD FJP.