Make GraphIndexBuilder.markNodeDeleted thread-safe
`GraphIndexBuilder.markNodeDeleted` uses `GrowableBitSet` as the backing storage
to track deleted nodes. But since `GrowableBitSet` itself is not thread-safe,
`markNodeDeleted` is not thread-safe either.

Added a synchronized version of `GrowableBitSet` and used it as the storage for
deleted nodes.

Currently `SynchronizedGrowableBitSet` uses a single lock to guard all access to `BitSet`.
If more concurrency needed while deleting nodes, a lock-free & atomic version can be introduced.

Closes #174