Remove on-disk reranking
On-disk reranking looks fast enough in Bench but that is because the datasets are small enough that the index under construction stays in the page cache, so reranking "from disk" is really reranking from page cache.

Performance drops off rapidly as the index outgrows the page cache.

Testing shows that the difference in accuracy between an index constructed with reranking and one constructed only with PQ is small and can be made up with minimal (~10%) more overquery.

Thus, it looks like this is not in fact a useful option and we are better off simplifying the code around LTM construction without it.