Handle unreachable nodes from the entry point better while removing deleted nodes
It is unlikely, but possible to have no path to some nodes from
the entry point of the graph. It might happen for example when you
have added a group of nodes that are very close to each other,
and then added another group of nodes that are very far away from
the first group such that there are no backlinks to these new nodes
from the initial group.

In that case, when you delete every node from the initial group,
you end up with some live nodes that are unreachable. So, we
set the entry point to -1 in removeDeletedNodes, although we have
some live nodes.

That breaks the logic in cleanup method, as we expect a valid
entry point when the graph is not empty. We recently hit a production
issue in a scenario what we believe was similar to this(we can't confirm
it fully as we can't access the customer data, but deploying a new version
with this fix solved the issue).

To handle this case, when the entry point is deleted in removeDeletedNodes,
we try to update the entry point as usual. If we can't do that
although there are some live nodes, we then assign a random entry
point to the graph.