Non-uniform vector quantization (#374)
Adds support for Non-uniform Vector Quantization (NVQ, pronounced as "new vec"). This new technique quantizes the values  in each vector with high accuracy by first applying a nonlinear transformation that is individually fit to each  vector. These nonlinearities are designed to be lightweight and have a negligible impact on distance computation  performance.

---------

Co-authored-by: Joel Knighton <joel.knighton@datastax.com>
Co-authored-by: Ted Willke <ted.willke@datastax.com>